+++ client/src/cli.rs
@@ -1,4 +1,4 @@
-use avail_light_core::network::Network;
+use avail_light_core::network::{Network, ServiceMode};
 use clap::{command, ArgAction, Parser};
 use tracing::Level;
 
@@ -73,4 +73,23 @@ pub struct CliOpts {
 	pub tracking_service_ping_interval: Option<u64>,
 	#[arg(long, default_value = "false")]
 	pub no_update: bool,
+
+	/// AutoNAT behaviour mode: client, server, or disabled
+	#[arg(long, value_enum, default_value = "server")]
+	pub auto_nat_mode: ServiceMode,
+
+	/// Relay behaviour mode: client, server, or disabled
+	#[arg(long, value_enum, default_value = "disabled")]
+	pub relay_mode: ServiceMode,
+
+	/// Enable or disable DCUTR (Direct Connection Upgrade through Relay)
+	/// Only works when relay_mode is Client or Both
+	#[arg(
+        long,
+        default_value = "false",
+        action = ArgAction::SetTrue,
+        requires = "relay_mode",
+        required_if_eq_any = [("relay_mode", "client"), ("relay_mode", "server")],
+    )]
+	pub enable_dcutr: bool,
 }